New Facilities - Now you can cancel splitting, archiving and extracting - Add archive status and extract status(not accurate yet) - Fix leechzipwatch if size more than 2gb - Other minor bugs fixed

+++ bot/modules/mirror.py
@@ -5,8 +5,7 @@ from time import sleep, time
 from os import path as ospath, remove as osremove, listdir, walk
 from shutil import rmtree
 from threading import Thread
-from subprocess import run as srun
-from pathlib import PurePath
+from subprocess import Popen
 from html import escape
 from telegram.ext import CommandHandler
 from telegram import InlineKeyboardMarkup
@@ -69,30 +68,34 @@ class MirrorListener:
             download = download_dict[self.uid]
             name = str(download.name()).replace('/', '')
             gid = download.gid()
-            size = download.size_raw()
             if name == "None" or self.isQbit or not ospath.exists(f'{DOWNLOAD_DIR}{self.uid}/{name}'):
                 name = listdir(f'{DOWNLOAD_DIR}{self.uid}')[-1]
             m_path = f'{DOWNLOAD_DIR}{self.uid}/{name}'
+        size = get_path_size(m_path)
         if self.isZip:
-            try:
-                with download_dict_lock:
-                    download_dict[self.uid] = ZipStatus(name, m_path, size)
-                path = m_path + ".zip"
-                LOGGER.info(f'Zip: orig_path: {m_path}, zip_path: {path}')
-                if self.pswd is not None:
-                    if self.isLeech and int(size) > TG_SPLIT_SIZE:
-                        srun(["7z", f"-v{TG_SPLIT_SIZE}b", "a", "-mx=0", f"-p{self.pswd}", path, m_path])
-                    else:
-                        srun(["7z", "a", "-mx=0", f"-p{self.pswd}", path, m_path])
-                elif self.isLeech and int(size) > TG_SPLIT_SIZE:
-                    srun(["7z", f"-v{TG_SPLIT_SIZE}b", "a", "-mx=0", path, m_path])
+            path = m_path + ".zip"
+            with download_dict_lock:
+                download_dict[self.uid] = ZipStatus(name, size, gid, self)
+            if self.pswd is not None:
+                if self.isLeech and int(size) > TG_SPLIT_SIZE:
+                    LOGGER.info(f'Zip: orig_path: {m_path}, zip_path: {path}.0*')
+                    self.arch_proc = Popen(["7z", f"-v{TG_SPLIT_SIZE}b", "a", "-mx=0", f"-p{self.pswd}", path, m_path])
                 else:
-                    srun(["7z", "a", "-mx=0", path, m_path])
-            except FileNotFoundError:
-                LOGGER.info('File to archive not found!')
-                self.onUploadError('Internal error occurred!!')
+                    LOGGER.info(f'Zip: orig_path: {m_path}, zip_path: {path}')
+                    self.arch_proc = Popen(["7z", "a", "-mx=0", f"-p{self.pswd}", path, m_path])
+            elif self.isLeech and int(size) > TG_SPLIT_SIZE:
+                LOGGER.info(f'Zip: orig_path: {m_path}, zip_path: {path}.0*')
+                self.arch_proc = Popen(["7z", f"-v{TG_SPLIT_SIZE}b", "a", "-mx=0", path, m_path])
+            else:
+                LOGGER.info(f'Zip: orig_path: {m_path}, zip_path: {path}')
+                self.arch_proc = Popen(["7z", "a", "-mx=0", path, m_path])
+            self.arch_proc.wait()
+            if self.arch_proc.returncode == -9:
                 return
-            if not self.isQbit or not self.seed or self.isLeech:
+            elif self.arch_proc.returncode != 0:
+                LOGGER.error('An error occurred while zipping! Uploading anyway')
+                path = f'{DOWNLOAD_DIR}{self.uid}/{name}'
+            if self.arch_proc.returncode == 0 and (not self.isQbit or not self.seed or self.isLeech):
                 try:
                     rmtree(m_path)
                 except:
@@ -103,7 +106,7 @@ class MirrorListener:
                     path = get_base_name(m_path)
                 LOGGER.info(f"Extracting: {name}")
                 with download_dict_lock:
-                    download_dict[self.uid] = ExtractStatus(name, m_path, size)
+                    download_dict[self.uid] = ExtractStatus(name, size, gid, self)
                 if ospath.isdir(m_path):
                     for dirpath, subdir, files in walk(m_path, topdown=False):
                         for file_ in files:
@@ -111,22 +114,30 @@ class MirrorListener:
                                or (file_.endswith(".rar") and not re_search(r'\.part\d+\.rar$', file_)):
                                 m_path = ospath.join(dirpath, file_)
                                 if self.pswd is not None:
-                                    result = srun(["7z", "x", f"-p{self.pswd}", m_path, f"-o{dirpath}", "-aot"])
+                                    self.ext_proc = Popen(["7z", "x", f"-p{self.pswd}", m_path, f"-o{dirpath}", "-aot"])
                                 else:
-                                    result = srun(["7z", "x", m_path, f"-o{dirpath}", "-aot"])
-                                if result.returncode != 0:
-                                    LOGGER.error('Unable to extract archive!')
-                        for file_ in files:
-                            if file_.endswith((".rar", ".zip", ".7z")) or re_search(r'\.r\d+$|\.7z\.\d+$|\.z\d+$|\.zip\.\d+$', file_):
-                                del_path = ospath.join(dirpath, file_)
-                                osremove(del_path)
+                                    self.ext_proc = Popen(["7z", "x", m_path, f"-o{dirpath}", "-aot"])
+                                self.arch_proc.wait()
+                                if self.ext_proc.returncode == -9:
+                                    return
+                                elif self.ext_proc.returncode != 0:
+                                    LOGGER.error('Unable to extract archive splits! Uploading anyway')
+                        if self.ext_proc.returncode == 0:
+                            for file_ in files:
+                                if file_.endswith((".rar", ".zip", ".7z")) or \
+                                    re_search(r'\.r\d+$|\.7z\.\d+$|\.z\d+$|\.zip\.\d+$', file_):
+                                    del_path = ospath.join(dirpath, file_)
+                                    osremove(del_path)
                     path = f'{DOWNLOAD_DIR}{self.uid}/{name}'
                 else:
                     if self.pswd is not None:
-                        result = srun(["bash", "pextract", m_path, self.pswd])
+                        self.ext_proc = Popen(["bash", "pextract", m_path, self.pswd])
                     else:
-                        result = srun(["bash", "extract", m_path])
-                    if result.returncode == 0:
+                        self.ext_proc = Popen(["bash", "extract", m_path])
+                    self.ext_proc.wait()
+                    if self.ext_proc.returncode == -9:
+                        return
+                    elif self.ext_proc.returncode == 0:
                         LOGGER.info(f"Extracted Path: {path}")
                         osremove(m_path)
                     else:
@@ -137,8 +148,7 @@ class MirrorListener:
                 path = f'{DOWNLOAD_DIR}{self.uid}/{name}'
         else:
             path = f'{DOWNLOAD_DIR}{self.uid}/{name}'
-        up_name = PurePath(path).name
-        up_path = f'{DOWNLOAD_DIR}{self.uid}/{up_name}'
+        up_name = path.rsplit('/', 1)[-1]
         if self.isLeech and not self.isZip:
             checked = False
             for dirpath, subdir, files in walk(f'{DOWNLOAD_DIR}{self.uid}', topdown=False):
@@ -149,9 +159,11 @@ class MirrorListener:
                         if not checked:
                             checked = True
                             with download_dict_lock:
-                                download_dict[self.uid] = SplitStatus(up_name, up_path, size)
+                                download_dict[self.uid] = SplitStatus(up_name, size, gid, self)
                             LOGGER.info(f"Splitting: {up_name}")
-                        split_file(f_path, f_size, file_, dirpath, TG_SPLIT_SIZE)
+                        res = split_file(f_path, f_size, file_, dirpath, TG_SPLIT_SIZE, self)
+                        if not res:
+                            return
                         osremove(f_path)
         if self.isLeech:
             size = get_path_size(f'{DOWNLOAD_DIR}{self.uid}')
@@ -163,6 +175,7 @@ class MirrorListener:
             update_all_messages()
             tg.upload()
         else:
+            up_path = f'{DOWNLOAD_DIR}{self.uid}/{up_name}'
             size = get_path_size(up_path)
             LOGGER.info(f"Upload Name: {up_name}")
             drive = GoogleDriveHelper(up_name, self)
